v0.8.2

+++ hyperdbg/hprdbghv/code/features/CompatibilityChecks.c
@@ -95,7 +95,7 @@ CompatibilityCheckModeBasedExecution()
     // The PML address and PML index fields exist only on processors that support the 1-setting of
     // the "enable PML" VM - execution control
     //
-    ULONG SecondaryProcBasedVmExecControls = HvAdjustControls(IA32_VMX_PROCBASED_CTLS2_MODE_BASED_EXECUTE_CONTROL_FOR_EPT_FLAG,
+    UINT32 SecondaryProcBasedVmExecControls = HvAdjustControls(IA32_VMX_PROCBASED_CTLS2_MODE_BASED_EXECUTE_CONTROL_FOR_EPT_FLAG,
                                                               IA32_VMX_PROCBASED_CTLS2);
 
     if (SecondaryProcBasedVmExecControls & IA32_VMX_PROCBASED_CTLS2_MODE_BASED_EXECUTE_CONTROL_FOR_EPT_FLAG)
@@ -126,7 +126,7 @@ CompatibilityCheckPml()
     // The PML address and PML index fields exist only on processors that support the 1-setting of
     // the "enable PML" VM - execution control
     //
-    ULONG SecondaryProcBasedVmExecControls = HvAdjustControls(IA32_VMX_PROCBASED_CTLS2_ENABLE_PML_FLAG, IA32_VMX_PROCBASED_CTLS2);
+    UINT32 SecondaryProcBasedVmExecControls = HvAdjustControls(IA32_VMX_PROCBASED_CTLS2_ENABLE_PML_FLAG, IA32_VMX_PROCBASED_CTLS2);
 
     if (SecondaryProcBasedVmExecControls & IA32_VMX_PROCBASED_CTLS2_ENABLE_PML_FLAG)
     {
@@ -145,7 +145,7 @@ CompatibilityCheckPml()
 }
 
 /**
- * @brief Checks for the compatiblity features based on current processor
+ * @brief Checks for the compatibility features based on current processor
  * @detail NOTE: NOT ALL OF THE CHECKS ARE PERFORMED HERE
  * @return VOID
  */
