+++ hyperdbg/hyperdbg-test/code/tests/lookup.cpp
@@ -42,7 +42,7 @@ TestListenerForResultsThread(void * data)
     BufferToRead = (char *)malloc(TEST_CASE_MAXIMUM_BUFFERS_TO_COMMUNICATE);
 
     //
-    // Wait for message from the target processs
+    // Wait for message from the target process
     //
     if (!NamedPipeServerWaitForClientConntection(OutputPipeHandle))
     {
@@ -182,7 +182,7 @@ TestCreateLookupTable(HANDLE PipeHandle, PVOID KernelInformation, UINT32 KernelI
 
         string OutputCommand = "cmd:" + NewCase;
 
-        SentMessageResult = NamedPipeClientSendMessage(PipeHandle, (char *)OutputCommand.c_str(), OutputCommand.length() + 1);
+        SentMessageResult = NamedPipeClientSendMessage(PipeHandle, (char *)OutputCommand.c_str(), (int)OutputCommand.length() + 1);
         if (!SentMessageResult)
         {
             //
